+++ l2-frontend/src/ui-cards/CashRegisters/ShiftModal.vue
